NamNS: add authorization for Sheet operations

+++ src/main/java/com/prc391/patra/sheets/SheetService.java
@@ -1,16 +1,18 @@
 package com.prc391.patra.sheets;
 
 import com.prc391.patra.constant.SecurityConstants;
+import com.prc391.patra.exceptions.EntityExistedException;
 import com.prc391.patra.exceptions.EntityNotFoundException;
 import com.prc391.patra.exceptions.UnauthorizedException;
+import com.prc391.patra.members.MemberRepository;
 import com.prc391.patra.tasks.Task;
 import com.prc391.patra.tasks.TaskRepository;
 import com.prc391.patra.utils.AuthorizationUtils;
 import lombok.AllArgsConstructor;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
+import java.util.Arrays;
 import java.util.List;
 import java.util.Optional;
 
@@ -20,7 +22,7 @@ public class SheetService {
     private final SheetRepository sheetRepository;
     private final TaskRepository taskRepository;
     private final AuthorizationUtils authorizationUtils;
-
+    private final MemberRepository memberRepository;
 
     public List<Sheet> getSheetFromOrgID(String orgID) throws EntityNotFoundException, UnauthorizedException {
         if (authorizationUtils.authorizeAccess(orgID, SecurityConstants.READ_ACCESS)) {
@@ -34,30 +36,59 @@ public class SheetService {
         }
     }
 
-    Sheet getSheetById(String sheetId) throws EntityNotFoundException {
+    Sheet getSheetById(String sheetId) throws EntityNotFoundException, UnauthorizedException {
         Optional<Sheet> result = sheetRepository.findById(sheetId);
-        if (result.isPresent()) {
-            return result.get();
-        } else {
-            throw new EntityNotFoundException();
+        if (!result.isPresent()) {
+            throw new EntityNotFoundException("Sheet " + sheetId + " not exist");
         }
+        Sheet sheet = result.get();
+        if (!authorizationUtils.authorizeAccess(sheet.getOrgId(), SecurityConstants.READ_ACCESS)) {
+            throw new UnauthorizedException("You don't have permission to access this resource");
+        }
+        return sheet;
     }
 
-    public List<Task> getTaskFromSheetId(String sheetId) throws EntityNotFoundException {
+    public List<Task> getTaskFromSheetId(String sheetId) throws EntityNotFoundException, UnauthorizedException {
         Optional<Sheet> result = sheetRepository.findById(sheetId);
         if (!result.isPresent()) {
             throw new EntityNotFoundException("Sheet with id " + sheetId + " is not exist!");
+        }
+        Sheet sheet = result.get();
+        if (!authorizationUtils.authorizeAccess(sheet.getOrgId(), SecurityConstants.READ_ACCESS)) {
+            throw new UnauthorizedException("You don't have permission to access this resource");
         }
         List<Task> taskList = taskRepository.getAllBySheetId(sheetId);
         return taskList;
     }
 
-    Sheet insertSheet(Sheet sheet) {
+    Sheet insertSheet(Sheet sheet) throws UnauthorizedException, EntityNotFoundException, EntityExistedException {
+        if (!authorizationUtils.authorizeAccess(sheet.getOrgId(), SecurityConstants.ADMIN_ACCESS)) {
+            throw new UnauthorizedException("You don't have permission to access this resource");
+        }
+        Optional<Sheet> optionalSheet = sheetRepository.findById(sheet.getSheetId());
+        if (optionalSheet.isPresent()) {
+            throw new EntityExistedException("Sheet " + sheet.getSheetId() + " is existed");
+        }
         return sheetRepository.save(sheet);
     }
 
-    void deleteSheet(String sheetId) {
-        //TODO check permission
+    void deleteSheet(String sheetId) throws EntityNotFoundException, UnauthorizedException, EntityExistedException {
+        Optional<Sheet> optionalSheet = sheetRepository.findById(sheetId);
+        if (optionalSheet.isPresent()) {
+            throw new EntityExistedException("Sheet " + sheetId + " is existed");
+        }
+        Sheet sheet = optionalSheet.get();
+        if (!authorizationUtils.authorizeAccess(sheet.getOrgId(), SecurityConstants.ADMIN_ACCESS)) {
+            throw new UnauthorizedException("You don't have permission to access this resource");
+        }
+        List<Task> taskInSheet = taskRepository.getAllBySheetId(sheetId);
+        for (Task task : taskInSheet) {
+            //remove assign information
+            List<String> oldAssigneeIds = task.getAssignee();
+            taskRepository.removeAssignee(task.getTaskId(), oldAssigneeIds);
+            memberRepository.removeAssignedTask(oldAssigneeIds, Arrays.asList(task.getTaskId()));
+        }
+        taskRepository.deleteAll(taskInSheet);
         sheetRepository.deleteById(sheetId);
     }
 }
